+++ RELEASE-NOTES.md
@@ -11,6 +11,8 @@
 
 ### API Changes
 
+1. Remove SQL formatting feature [#35390](https://github.com/apache/shardingsphere/pull/35390)
+
 ### New Features
 
 ### Enhancements
@@ -35,6 +37,7 @@
 1. SQL Parser: Support parsing MySQL stored procedure syntax - [#35137](https://github.com/apache/shardingsphere/pull/35137)
 1. SQL Parser: Support Oracle SQL parsing V1 keywords as identifiers - [#35373](https://github.com/apache/shardingsphere/pull/35373)
 1. SQL Parser: Support Oracle in literal sql parsing - [#35384](https://github.com/apache/shardingsphere/pull/35384)
+1. SQL Bind: Support explain statement sql bind - [#35439](https://github.com/apache/shardingsphere/pull/35439)
 
 ### Bug Fixes